Add remote::query module
Provides common HTTP calls and handling for remote
services. This is the case where Gitlab and Github
provide similar behavior and we can reuse the
code.
